requirements.txt
discord.py==1.6.0
requests==2.25.1
giphypop==0.3
TenGiphPy==2.5.0